refactor(textarea): split into 8 focused submodules
- Extract content to content.rs (202 lines)
- Extract editing to editing.rs (219 lines)
- Extract find_impl to find_impl.rs (346 lines)
- Extract multi_cursor to multi_cursor.rs (153 lines)
- Extract navigation to navigation.rs (171 lines)
- Extract undo to undo.rs (106 lines)
- Extract view to view.rs (193 lines)
- Original mod.rs was 1,382 lines
